==> 00-h5-to-waterindex.Rmd <==
---
title: "centerlines"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Find river centerline from a surface reflectance file

```{r}
library(terra)
library(hdf5r)
library(dplyr)
library(sf)
library(glue)
library(fs)
library(stringr)
library(stars)
library(scico)
library(ggplot2)
```

# Set up

```{r}
my_aop_yr <- '2020'
my_aop_site <- 'NOGP'
my_domain <- 'D09'
```

get all L3 files for site

```{r}
my_site_dir <- glue('D:/{my_aop_yr}/FullSite/{my_domain}') %>%
  fs::dir_ls(glob = glue('*{my_aop_site}*'), type = 'directory')
my_site_files <- glue('{my_site_dir}/L3/Spectrometer/Reflectance') %>%
  fs::dir_ls(glob = '*.h5', recurse = TRUE)
length(my_site_files)
```


```{r}
# my_h5_file <- my_site_files[2]
```


# For each file

```{r}

check_waterindex <- function(my_h5_file){
  my_h5 <- hdf5r::H5File$new(my_h5_file, mode = "r")
  # reflectance and metadata
  epsg_path <- glue::glue('{my_aop_site}/Reflectance/Metadata/Coordinate_System/EPSG Code')
  my_epsg <- my_h5[[epsg_path]]$read() %>% as.integer()
  my_epsg2 <- glue('EPSG:{my_epsg}')
  #
  my_flightline_id <- my_h5_file %>% stringr::str_split('/') %>% unlist()
  my_reflectance_id <- tools::file_path_sans_ext(my_flightline_id)[9] # DEPENDS ON FOLDER WHERE H5 FILE IS
  #
  my_refl <- my_h5[[glue('{my_aop_site}/Reflectance/Reflectance_Data')]]
  wls_path <- glue('{my_aop_site}/Reflectance/Metadata/Spectral_Data/Wavelength')
  my_wls <- my_h5[[wls_path]]$read()
  scale_factor <- my_refl$attr_open('Scale_Factor')$read()
  na_value <- my_refl$attr_open('Data_Ignore_Value')$read()
    
  # WAVELENGTHS FOR ALGORITHMS
  wl560id <- which.min(abs(my_wls-560))
  wl1640id <- which.min(abs(my_wls-1640))
  wl485id <- which.min(abs(my_wls-485))
  wl830id <- which.min(abs(my_wls-830))
  wl1650id <- which.min(abs(my_wls-1650))
  wl2215id <- which.min(abs(my_wls-2215))
    
  map_info_path <- glue('{my_aop_site}/Reflectance/Metadata/Coordinate_System/Map_Info')
  map_info <- my_h5[[map_info_path]]$read() %>% strsplit(',') %>% unlist()
  my_dims <- my_refl$dims
  xy_resolution <- as.numeric(c(map_info[2], map_info[3]))
  xmin <- as.numeric(map_info[4])
  xmax <- xmin + my_dims[2] * xy_resolution[1]
  ymax <- as.numeric(map_info[5])
  ymin <- ymax - my_dims[3] * xy_resolution[2]
  my_extent <- terra::ext(xmin, xmax, ymin, ymax)
  
    
  my_bands <- c(wl560id = wl560id,
                wl485id = wl485id,
                wl830id = wl830id,
                wl1650id = wl1650id,
                wl2215id = wl2215id,
                  wl1640id = wl1640id) # select band IDs to extract
  
  my_spectra_list <- purrr::map(my_bands,
                                  ~terra::rast(my_refl[.x, 1:my_dims[2], 1:my_dims[3]]))
  my_spectra_list <- purrr::map(my_spectra_list, ~terra::t(.))
    
  my_rast_n <- terra::rast(my_spectra_list)
  my_rast_n[my_rast_n == -9999] <- NA
  names(my_rast_n) <- glue('b{stringr::str_pad(my_bands, 3, "left", "0")}_{round(my_wls[my_bands])}nm')
  
  terra::ext(my_rast_n) <- my_extent
  terra::crs(my_rast_n) <- my_epsg2
    
    
    # MODIFIED nNDWI (green - swir)/(green + swir)
    calc_mNDWI <- function(img, b1x, b2x){
      b1 <- img[[b1x]]
      b2 <- img[[b2x]]
      mndwi <- (b1 - b2)/(b1 + b2)
      return(mndwi)
    }
    b560 <- grep('wl560id', names(my_bands))
    b1640 <- grep('wl1640id', names(my_bands))
    my_mndwi <- calc_mNDWI(my_rast_n, b560, b1640)
    names(my_mndwi) <- 'mNDWI'
    
    # AWEI Feyisa et al 2014
    calc_AWEI <- function(img, b1x, b2x, b4x, b5x, b7x){
      b1 <- img[[b1x]]
      b2 <- img[[b2x]]
      b4 <- img[[b4x]]
      b5 <- img[[b5x]]
      b7 <- img[[b7x]]
      awei2 <- 4*(b2 - b5) - (0.25*b4 + 2.75*b7)
      awei3 <- b1 + (2.5*b2) - (1.5*(b4 + b5)) - (0.25*b7)
      awei <- awei2 > 0 & awei3 > 0
      return(awei)
    }
    b485 <- grep('wl485id', names(my_bands))
    b560 <- grep('wl560id', names(my_bands))
    b830 <- grep('wl830id', names(my_bands))
    b1650 <- grep('wl1650id', names(my_bands))
    b2215 <- grep('wl2215id', names(my_bands))
    my_awei <- calc_AWEI(my_rast_n,b485, b560, b830, b1650, b2215)
    names(my_awei) <- 'AWEI'
    
    my_mndwi_thresh <- my_mndwi > 0
    my_wi <- my_awei - my_mndwi_thresh

    hdf5r::h5close(my_h5)
  
  any_water_check <- max(terra::minmax(my_awei)) > 0 # check if there is any water in image!
  if(any_water_check){
    data_dir <- 'S:/users/khondula/DATA'
    mndwi_dir <- glue('{data_dir}/centerlines/01_mNDWI/{basename(my_site_dir)}')
    fs::dir_create(mndwi_dir)
    my_w1 <- my_mndwi %>% terra::wrap()
    saveRDS(my_w1, file = glue('{mndwi_dir}/{my_reflectance_id}.rds'))
    my_filename1 <- glue('{mndwi_dir}/{my_reflectance_id}.tif')
    terra::writeRaster(x = my_mndwi, filename = my_filename1, overwrite = TRUE)
    
    awei_dir <- glue('{data_dir}/centerlines/01_AWEI/{basename(my_site_dir)}')
    fs::dir_create(awei_dir)
    my_w2 <- my_awei %>% terra::wrap()
    saveRDS(my_w2, file = glue('{awei_dir}/{my_reflectance_id}.rds'))
    my_filename2 <- glue('{awei_dir}/{my_reflectance_id}.tif')
    terra::writeRaster(x = my_awei, filename = my_filename2, overwrite = TRUE)

        
    cam_dir <- my_h5_file %>% fs::path_dir() %>% fs::path_dir() %>% path_dir()
    cam_file <- fs::dir_ls(glue('{cam_dir}/Camera/Mosaic'), glob = glue('*{xmin}_{ymin}*'))
    check_cam_file <- length(cam_file)>0
    if(check_cam_file){
      my_cam <- terra::rast(cam_file)
      plot(my_cam, main = 'camera')
      cc <- recordPlot()
    }
    if(!check_cam_file){cc = NA}
    
    my_bbox_sf <- my_awei %>% st_bbox() %>% st_as_sfc() %>% st_as_sf()
    
    aa <- ggplot() +
      geom_sf(data = my_bbox_sf) +
      geom_stars(data = st_as_stars(my_awei)) +
      colorspace::scale_fill_continuous_diverging(palette = 'Berlin', rev = TRUE) +
      theme_void() +
      # ggtitle('AWEI') +
      # ggtitle(glue('{my_reflectance_id}')) +
      theme(legend.position = 'none')
    bb <- ggplot() +
      geom_sf(data = my_bbox_sf) +
      geom_stars(data = st_as_stars(my_mndwi)) +
      colorspace::scale_fill_continuous_diverging(palette = 'Cork', mid = 0, rev = TRUE) +
      theme_void() +
      # ggtitle('mNDWI') +
      # ggtitle(glue('{my_reflectance_id}')) +
      theme(legend.position = 'none')
    
    dd <- ggplot() +
      geom_sf(data = my_bbox_sf) +
      geom_stars(data = st_as_stars(my_wi)) +
      colorspace::scale_fill_continuous_diverging(palette = 'Berlin', mid = 0, rev = TRUE) +
      theme_void() +
      # ggtitle('mNDWI') +
      # ggtitle(glue('{my_reflectance_id}')) +
      theme(legend.position = 'none')
  
  wm2 <- cowplot::plot_grid(cc, aa, bb, dd, nrow = 1,
                            rel_widths = c(0.23, 0.25, 0.25, 0.25),
                            labels = c("camera", "AWEI", "MNDWI", "AWEI - MNDWI"))
  plot_dir <- glue('{data_dir}/centerlines/01_watermask-imgs/{basename(my_site_dir)}')
  fs::dir_create(plot_dir)
  my_plotfile <- glue::glue('{plot_dir}/{my_reflectance_id}.png')
  ggsave(my_plotfile, wm2, width = 12, height = 4)
 
  }
  message(glue('saved {my_reflectance_id} watermask'))  
}

```

```{r}
my_site_files %>% purrr::walk(~check_waterindex(.x))
```

# delete files

```{r}
# my_aop_yr <- '2021'
# my_aop_site <- 'SCBI'
# my_domain <- 'D02'
# my_site_dir <- glue('D:/{my_aop_yr}/FullSite/{my_domain}') %>%
#   fs::dir_ls(glob = glue('*{my_aop_site}*'), type = 'directory')
# my_site_files <- glue('{my_site_dir}/L3/Spectrometer/Reflectance') %>%
#   fs::dir_ls(glob = '*.h5', recurse = TRUE)
my_site_files %>% purrr::walk(~fs::file_delete(.x))

my_site_dir <- glue('D:/{my_aop_yr}/FullSite/{my_domain}') %>%
  fs::dir_ls(glob = glue('*{my_aop_site}*'), type = 'directory')
my_site_files <- glue('{my_site_dir}/L3/Camera/Mosaic') %>%
  fs::dir_ls(glob = '*.tif', recurse = TRUE)
my_site_files %>% purrr::walk(~fs::file_delete(.x))
```
