==> 01-mNDWI-imgs.Rmd <==
---
title: "centerlines"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Find river centerline from a surface reflectance file

```{r}
library(terra)
library(hdf5r)
library(dplyr)
library(sf)
library(glue)
library(stringr)
library(stars)
library(scico)
library(ggplot2)
library(smoothr)
# library(gdistance)
# library(leaflet)
# library(whitebox)

```

# whitebox part

* make sure connected to scratch

```{r}
# data_dir <- '/Volumes/AOPScratch/users/khondula/DATA'
data_dir <- 'S:/users/khondula/DATA'
fs::dir_ls(glue('{data_dir}/centerlines/01_mNDWI')) %>% basename()
```

```{r}
my_site_yr <- '2021_SERC_5'
my_aop_site <- 'SERC'
my_aop_yr <- 2021
```

```{r}
my_site_dir <- glue('{data_dir}/centerlines/01_mNDWI/{my_site_yr}')
my_site_files <- fs::dir_ls(my_site_dir)

```

map with nhd flowlines also? 

```{r}
# my_site_file <- my_site_files[17]

save_watermask <- function(my_site_file){
  my_mndwi <- readRDS(file = my_site_file) %>% terra::rast()
  my_reflectance_id <- basename(my_site_file) %>% tools::file_path_sans_ext()
  
  aa <- ggplot() +
      geom_stars(data = st_as_stars(my_mndwi)) +
      scico::scale_fill_scico(palette = 'oleron', direction = -1) +
      theme_void() +
      ggtitle(glue('{my_reflectance_id}')) +
      theme(legend.position = 'bottom')
  
  # threshold for water mask
  rcl_matrix <- matrix(c(-Inf, 0, 0, 
                           0, Inf, 1), ncol = 3, byrow = TRUE)
  riv_mask <- terra::classify(my_mndwi, rcl = rcl_matrix, include.lowest = FALSE)
  
  bb <- ggplot() +
      geom_stars(data = st_as_stars(riv_mask)) +
      theme_void() +
      ggtitle(glue('{my_reflectance_id}')) +
      theme(legend.position = 'bottom')
  
  wm <- cowplot::plot_grid(aa, bb, nrow = 1)
  
  my_wm_dir <- glue('{data_dir}/centerlines/02_watermask-imgs/{my_site_yr}')
  fs::dir_create(my_wm_dir)
  my_wm_filename <- glue('{my_wm_dir}/{my_reflectance_id}.png')
  ggsave(my_wm_filename, wm)
  message(glue('saved {my_reflectance_id} watermask'))
}


```

make QA spreadsheet - do this while it is fresh!

```{r}
df <- data.frame(refl_id = basename(my_site_files)) %>%
  dplyr::mutate(site_yr_id = my_site_yr) %>%
  dplyr::mutate(refl_id = tools::file_path_sans_ext(refl_id)) %>%
  dplyr::mutate(river_check = NA) %>%
  dplyr::select(site_yr_id, refl_id, river_check)

df %>% readr::write_excel_csv(file = glue('{my_site_yr}-rivercheck.csv'), na = "")

# df %>% readr::write_excel_csv(file = glue('~/Box/data/watermask-check/{my_site_yr}-rivercheck.csv'), na = "")
```


```{r}
my_site_files %>% purrr::walk(~save_watermask(.x))
```

# done

# after checking water masks

```{r}
checked_site_yr <- '2019_TOOL_3'
check_df <- readr::read_csv(glue('{checked_site_yr}-rivercheck-notes.csv')) %>% dplyr::filter(river_check)
# check_df <- fs::dir_ls('~/Box/data/watermask-check/done', glob = glue('*{my_site_yr}*')) %>%
#   readr::read_csv() %>%
#   dplyr::filter(river_check)
my_river_fileids <- check_df$refl_id
check_df %>% readr::write_csv(glue('{data_dir}/riverline-checks/{checked_site_yr}.csv'))
```

```{r}
check_df <- readr::read_csv(glue('{data_dir}/riverline-checks/{checked_site_yr}.csv'))
my_river_fileids <- check_df$refl_id
my_site_dir <- glue('{data_dir}/centerlines/01_mNDWI/{checked_site_yr}')
my_site_files <- fs::dir_ls(my_site_dir)
idz <- my_river_fileids %>% purrr::map_int(~grep(pattern = .x, x = my_site_files))
river_site_files <- my_site_files[idz]
```

```{r}
rlist <- river_site_files %>%
  purrr::map(~readRDS(.x)) %>%
  purrr::map(~terra::rast(.x))
rsrc <- terra::src(rlist)
m <- terra::mosaic(rsrc)
wrap(m) %>% saveRDS(glue('{data_dir}/centerlines/00_mosaic-data/{checked_site_yr}-mosaic.rds'))
# plot(m)

mm <- ggplot() +
    geom_stars(data = st_as_stars(m)) +
    scico::scale_fill_scico(palette = 'oleron', direction = -1) +
    theme_void() +
    theme(legend.position = 'bottom')
# ggsave(filename = glue('{data_dir}/centerlines/00_mosaic-data/{my_site_yr}-mosaic.png'), mm)
ggsave(filename = glue('{data_dir}/centerlines/00_mosaic-data/{checked_site_yr}-mosaic.pdf'), mm, width = 8, height = 10)
```

```{r}
river_site_file <- river_site_files[1]
my_mndwi <- m
```

```{r}
  my_mndwi <- readRDS(file = river_site_file) %>% terra::rast()
  my_reflectance_id <- basename(river_site_file) %>% tools::file_path_sans_ext()
  
  # threshold for water mask
  rcl_matrix <- matrix(c(-Inf, 0, 0, 
                           0, Inf, 1), ncol = 3, byrow = TRUE)
  riv_mask <- terra::classify(my_mndwi, rcl = rcl_matrix, include.lowest = FALSE)
  # max filter for noise/tree canopies - widens the stream a bit
  riv_mask1 = focal(riv_mask, 
                   w = matrix(1, nrow = 3, ncol = 3), 
                   fun = max)

  riv_polygon <- riv_mask1 %>% terra::as.polygons()
  riv_polygon_sf <- riv_polygon %>% st_as_sf() %>% dplyr::filter(mNDWI == 1)

  ggplot() +
    geom_stars(data = st_as_stars(riv_mask1)) +
    geom_sf(data = riv_polygon_sf, fill = NA, col = 'yellow') +
    theme_void() +
   theme(legend.position = 'bottom')
 
  # SMOOTH with ksmooth 20
  riv_smooth <- riv_polygon_sf %>% smoothr::smooth(method = "ksmooth", smoothness = 20)

  # mask raster with smoothed river

  ggplot() +
    geom_stars(data = st_as_stars(riv_mask1)) +
    geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
    theme_void() +
   theme(legend.position = 'bottom')

  # make actual NAs as 0 before doing inverse!
  riv_mask1[is.na(riv_mask1)] <- 0
  
  riv_smooth_sv <- as(riv_smooth, "SpatVector")
  riv_mask2 <- terra::mask(riv_mask1, mask = riv_smooth_sv, inverse = TRUE, touches = TRUE)

  ggplot() +
    geom_stars(data = st_as_stars(riv_mask2)) +
    geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
    theme_void() +
    theme(legend.position = 'bottom')

```

```{r}
leaflet() %>%
  addProviderTiles(providers$Esri.WorldImagery) %>%
  addPolygons(data = st_transform(riv_smooth, 4326))
```

```{r}
my_reflectance_id <- 'blan_mosaic'
  # reclassify for whitebox
  rcl_matrix <- matrix(c(-Inf, 1, 1, 
                         NA, NA, 0), ncol = 3, byrow = TRUE)
  riv_mask_wb <- terra::classify(riv_mask2, rcl = rcl_matrix, include.lowest = FALSE)
  plot(riv_mask_wb)
  wb_filename1 <- glue('{my_reflectance_id}-rivmask_wb.tif')
  riv_mask_wb %>% writeRaster(wb_filename1)

    wb_filename2 <- glue('{my_reflectance_id}-wb_dists.tif')

wbt_euclidean_distance(
  input = wb_filename1,
  output = wb_filename2,
  verbose_mode = TRUE)
riv_dist <- terra::rast(wb_filename2)
plot(riv_dist)
```

```{r}

ggplot() +
    geom_stars(data = st_as_stars(riv_dist)) +
    geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
    theme_void() +
    theme(legend.position = 'bottom')

# edge detection with sobel

w_sobel1 <- matrix(c(-1, -2, -1, 0, 0, 0, 1, 2, 1), nrow = 3)
w_sobel2 <- matrix(c(1, 2, 1, 0, 0, 0, -1, -2, -1), byrow = TRUE, nrow = 3)

riv_dists_sobel1 = terra::focal(riv_dist, 
                   w = w_sobel1, 
                   expand = TRUE)
riv_dists_sobel2 = terra::focal(riv_dist, 
                   w = w_sobel2, 
                   expand = TRUE)

riv_sobel <- sqrt(riv_dists_sobel1^2 + riv_dists_sobel2^2)

ggplot() +
    geom_stars(data = st_as_stars(riv_sobel)) +
    # geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
    theme_void()

# remask with river
riv_sobel_mask <- terra::mask(riv_sobel, mask = riv_smooth_sv, touches = FALSE)

ggplot() +
    geom_stars(data = st_as_stars(riv_sobel_mask)) +
    geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
  theme_void() 



polygons_dir <-glue('polygons/{my_aop_site}_{my_aop_yr}')
fs::dir_create(polygons_dir)
polygon_filename <- glue('{polygons_dir}/{my_reflectance_id}-riversmooth.shp')
riv_smooth %>% 
  dplyr::mutate(aop_site = my_aop_site, aop_yr = my_aop_yr, refl_id = my_reflectance_id) %>% 
  st_write(polygon_filename, append = FALSE)
```

```{r}
# create points to manually identify river segments

# low threshold for zero crossings
my_val <- quantile(terra::values(riv_sobel_mask), 0.01, na.rm = TRUE) # 0.2 for area, 0.01 for points? 

rcl_matrix_pts <- matrix(c(-Inf, my_val, 1, 
                         my_val, Inf, NA), ncol = 3, byrow = TRUE)

riv_sobel_mask01 <- terra::classify(riv_sobel_mask, rcl = rcl_matrix_pts, include.lowest = FALSE)

# will be barely visible
ggplot() +
  geom_stars(data = st_as_stars(riv_sobel_mask01)) +
  theme_void() 

# convert to points
riv_cent_sv <- terra::as.points(riv_sobel_mask01)
riv_cent_sf <- riv_cent_sv %>% st_as_sf() %>% tibble::rowid_to_column('pt_id')
plot(riv_cent_sf)

```
 
# helper map for point ID

```{r}
my_reflectance_id
my_aop_yr <- 2021
```


```{r}
leaflet() %>%
  addTiles(group = 'bg') %>%
  addProviderTiles(providers$Esri.WorldImagery, group = 'esri') %>%
  addPolygons(data = st_transform(riv_smooth, 4326), color = 'cyan', fillOpacity = 0.1) %>%
  addMarkers(data = st_transform(riv_cent_sf, 4326),   
             clusterOptions = markerClusterOptions(),
             group = 'pts', popup = ~glue('point {pt_id}')) %>%
  addLayersControl(overlayGroups = c('bg','pts', 'esri'))
```

```{r}
bb <- ggplot() +
    geom_stars(data = st_as_stars(my_mndwi)) +
       colorspace::scale_fill_continuous_diverging(palette = 'Cork', mid = 0, rev = TRUE) +
     geom_sf(data = riv_smooth, fill = NA, col = NA) +
    theme_void() +
   theme(legend.position = 'bottom') +
  ggtitle(glue('{my_reflectance_id} {my_aop_yr}'))

cc <- ggplot() +
    geom_stars(data = st_as_stars(my_mndwi)) +
       colorspace::scale_fill_continuous_diverging(palette = 'Cork', mid = 0, rev = TRUE) +
     geom_sf(data = riv_smooth, fill = NA, col = 'black') +
    theme_void() +
   theme(legend.position = 'bottom') +
   ggtitle('')

cowplot::plot_grid(bb, cc, nrow = 1)

ggsave(glue('figs/{my_reflectance_id}-{my_aop_yr}.png'), width = 12)

```
