==> 02-watermask-mosaic.Rmd <==
---
title: "centerlines"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Find river centerline from a surface reflectance file

```{r}
library(terra)
library(hdf5r)
library(dplyr)
library(sf)
library(glue)
library(stringr)
library(stars)
library(scico)
library(ggplot2)
library(smoothr)
# library(gdistance)
# library(leaflet)
library(whitebox)

```


# after checking water masks

* make sure connected to scratch

```{r}
data_dir <- '/Volumes/AOPScratch/users/khondula/DATA'
# data_dir <- 'S:/users/khondula/DATA'
fs::dir_ls(glue('{data_dir}/centerlines/03_watermask-data')) %>% basename()
fs::dir_ls(glue('{data_dir}/centerlines/02_watermask-imgs')) %>% basename()
```

```{r}
# checked_site_yr <- '2021_BLAN_4'
# checked_site_yr <- '2021_NOGP_5'
checked_site_yr <- '2021_ORNL_5'
rivercheck_dir <- glue('{data_dir}/centerlines/riverchecks')
check_df <- readr::read_csv(glue('{rivercheck_dir}/{checked_site_yr}-rivercheck-notes.csv')) %>% dplyr::filter(river_check)
my_river_fileids <- check_df$refl_id
my_site_dir <- glue('{data_dir}/centerlines/01_mNDWI/{checked_site_yr}')
my_site_files <- fs::dir_ls(my_site_dir)
idz <- my_river_fileids %>% purrr::map_int(~grep(pattern = .x, x = my_site_files))
river_site_files <- my_site_files[idz]
```

classify 

```{r}
# river_site_file <- river_site_files[1]

save_watermask01 <- function(river_site_file){
  mndwi_rast <- river_site_file %>% readRDS() %>% terra::rast()
  my_reflectance_id <- basename(river_site_file) %>% tools::file_path_sans_ext()

  # threshold for water mask
  rcl_matrix <- matrix(c(-Inf, 0, 0, 
                           0, Inf, 1), ncol = 3, byrow = TRUE)
    riv_mask <- terra::classify(mndwi_rast, rcl = rcl_matrix, include.lowest = FALSE)

    # max filter for noise/tree canopies - widens the stream a bit
    riv_mask1 = terra::focal(riv_mask, 
                   w = matrix(1, nrow = 3, ncol = 3), 
                   expand = TRUE,
                   fun = max)

  # c(riv_mask, riv_mask1) %>% plot()

    mask_dir <- glue('{data_dir}/centerlines/03_watermask-data/{basename(my_site_dir)}')
    fs::dir_create(mask_dir)
    my_w <- riv_mask1 %>% terra::wrap()
    my_filename <- glue('{mask_dir}/{my_reflectance_id}.tif')
    terra::writeRaster(x = riv_mask1, filename = my_filename, overwrite = TRUE)
    saveRDS(my_w, file = glue('{mask_dir}/{my_reflectance_id}.rds'))
    message(glue('saved {my_reflectance_id} watermask'))  
}


```


```{r}
river_site_files %>% purrr::walk(~save_watermask01(.x))
```

## moasic

```{r}

mask_dir <- glue('{data_dir}/centerlines/03_watermask-data/{basename(my_site_dir)}')
# mask_dir <- glue('03_watermask-data/{basename(my_site_dir)}')
# mask_files <- fs::dir_ls(mask_dir, glob = glue('*.rds'))
mask_files <- fs::dir_ls(mask_dir, glob = glue('*.tif')) %>% as.character()

mosaic_outname <- glue('mosaic-{basename(my_site_dir)}.tif') %>% as.character()

wbt_mosaic(output = mosaic_outname,
           inputs = mask_files[1:2],
           verbose_mode = TRUE,
           compress_rasters = TRUE)

rlist <- mask_files %>%
  # purrr::map(~readRDS(.x)) %>%
  purrr::map(~terra::rast(.x)) %>%
  purrr::map(~terra::trim(.x))
rsrc <- terra::src(rlist)
m <- terra::mosaic(rsrc)
wrap(m) %>% saveRDS(glue('{data_dir}/centerlines/04_watermask-mosaic/{checked_site_yr}-mosaic.rds'))

m <- terra::rast('03_watermask-data/2021_ORNL_5/mosaic.tif')
plot(mm)
m <- mm
```

```{r}
# mm <- ggplot() +
#     geom_stars(data = st_as_stars(m)) +
#     # scico::scale_fill_scico(palette = 'oleron', direction = -1) +
#     theme_void() +
#     theme(legend.position = 'bottom') +
#   ggtitle(glue('{checked_site_yr} mNDWI > 0 w max filter'))
# ggsave(filename = glue('{data_dir}/centerlines/04_watermask-mosaic/{checked_site_yr}-mosaic.pdf'), mm, width = 12, height = 8)
```

# polygons

```{r}
riv_polygon <- m %>% terra::as.polygons()
riv_polygon_sf <- riv_polygon %>% st_as_sf() %>% dplyr::filter(mosaic == 1)
# SMOOTH with ksmooth 20
riv_smooth <- riv_polygon_sf %>% smoothr::smooth(method = "ksmooth", smoothness = 20)

# ggplot() +
#   geom_sf(data = riv_smooth)
# 
# polygons_dir <-glue('polygons/{my_aop_site}_{my_aop_yr}')
# fs::dir_create(polygons_dir)
# polygon_filename <- glue('{polygons_dir}/{my_reflectance_id}-riversmooth.shp')
# riv_smooth %>% 
#   dplyr::mutate(aop_site = my_aop_site, aop_yr = my_aop_yr, refl_id = my_reflectance_id) %>% 
#   st_write(polygon_filename, append = FALSE)
```

```{r}
# mask with smooth
# make actual NAs as 0 before doing inverse!
# m[is.na(m)] <- 0
# plot(m)  

riv_smooth_sv <- as(riv_smooth, "SpatVector")
riv_mask2 <- terra::mask(m, mask = riv_smooth_sv, inverse = TRUE, touches = TRUE, updatevalue = 1)
plot(riv_mask2)

rcl_matrix <- matrix(c(0, 1, 
                         1, 0), ncol = 2, byrow = TRUE)
# rcl_matrix <- matrix(c(-Inf, 1, 1, 
#                          NA, NA, 0), ncol = 3, byrow = TRUE)
riv_mask_wb <- terra::classify(riv_mask2, rcl = rcl_matrix)

# plot(riv_mask_wb)


# wb_dir <- glue('{data_dir}/centerlines/06_rivmask_4wb/{basename(my_site_dir)}')
# fs::dir_create(wb_dir)

wb_dir <- glue('06_rivmask_4wb/{basename(my_site_dir)}')
fs::dir_create(wb_dir)

wb_filename1 <- glue('{wb_dir}/rivmask_wb.tif')
riv_mask_wb %>% terra::writeRaster(wb_filename1, overwrite = TRUE)


```

# whitebox part

```{r}
wb_dir <- glue('06_rivmask_4wb/{basename(my_site_dir)}')
fs::dir_create(wb_dir)

wb_filename1 <- glue('{wb_dir}/rivmask_wb.tif')
# wb_dir2 <- glue('{data_dir}/centerlines/07_wb-dist/{basename(my_site_dir)}')
# fs::dir_create(wb_dir2)
wb_dir2 <- glue('07_wb-dist/{basename(my_site_dir)}')
fs::dir_create(wb_dir2)
wb_filename2 <- glue('{wb_dir2}/wb_dists.tif')

wbt_euclidean_distance(
  input = wb_filename1,
  output = wb_filename2,
  compress_rasters = TRUE,
  verbose_mode = TRUE)

riv_dist <- terra::rast(wb_filename2)
wb_dir2c <- glue('{data_dir}/centerlines/07_wb-dist/{basename(my_site_dir)}')
fs::dir_create(wb_dir2c)
riv_dist %>% terra::wrap() %>% saveRDS(glue('{wb_dir2c}/wb_dists.rds'))
riv_dist %>% terra::writeRaster(glue('{wb_dir2c}/wb_dists.tif'))

plot(riv_dist)

# mm2 <- ggplot() +
#     geom_stars(data = st_as_stars(riv_mask2)) +
#     # scico::scale_fill_scico(palette = 'oleron', direction = -1) +
#     theme_void() +
#     theme(legend.position = 'bottom') +
#   ggtitle(glue('{checked_site_yr} mNDWI > 0 w max filter'))
# ggsave(filename = glue('{data_dir}/centerlines/{checked_site_yr}-mosaic.png'), mm2)
```



```{r}

# ggplot() +
#     geom_stars(data = st_as_stars(riv_dist)) +
#     geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
#     theme_void() +
#     theme(legend.position = 'bottom')

# edge detection with sobel

w_sobel1 <- matrix(c(-1, -2, -1, 0, 0, 0, 1, 2, 1), nrow = 3)
w_sobel2 <- matrix(c(1, 2, 1, 0, 0, 0, -1, -2, -1), byrow = TRUE, nrow = 3)

riv_dists_sobel1 = terra::focal(riv_dist, 
                   w = w_sobel1, 
                   expand = TRUE)
riv_dists_sobel2 = terra::focal(riv_dist, 
                   w = w_sobel2, 
                   expand = TRUE)

riv_sobel <- sqrt(riv_dists_sobel1^2 + riv_dists_sobel2^2)

# ggplot() +
#     geom_stars(data = st_as_stars(riv_sobel)) +
#     geom_sf(data = riv_smooth, fill = NA, col = NA) +
#     theme_void()

# remask with river so areas outside are NA
riv_sobel_mask <- terra::mask(riv_sobel, mask = riv_smooth_sv, touches = FALSE)

# plot(riv_sobel_mask)
# gg <- ggplot() +
#     geom_stars(data = st_as_stars(riv_sobel_mask)) +
#     # geom_sf(data = riv_smooth, fill = NA, col = 'yellow') +
#   theme_void() 
# ggsave('test2.pdf', gg)
```

```{r}
# create points to manually identify river segments

# low threshold for zero crossings
my_val <- quantile(terra::values(riv_sobel_mask), 0.01, na.rm = TRUE) # 0.2 for area, 0.01 for points? 

rcl_matrix_pts <- matrix(c(-Inf, my_val, 1, 
                         my_val, Inf, NA), ncol = 3, byrow = TRUE)

riv_sobel_mask01 <- terra::classify(riv_sobel_mask, rcl = rcl_matrix_pts, include.lowest = FALSE)

# now higher threshold for area
my_val2 <- quantile(terra::values(riv_sobel_mask), 0.2, na.rm = TRUE)

rcl_matrix_area <- matrix(c(-Inf, my_val2, 1, 
                         my_val2, Inf, NA), ncol = 3, byrow = TRUE)

riv_sobel_mask20 <- terra::classify(riv_sobel_mask, rcl = rcl_matrix_area, include.lowest = FALSE)

# # will be barely visible
# gg2 <- ggplot() +
#   geom_stars(data = st_as_stars(riv_sobel_mask20)) +
#   theme_void() 
# 
# ggsave('test3.pdf', gg2)

# wbt_clump(
#   input,
#   output,
#   diag = TRUE,
#   zero_back = FALSE,
#   wd = NULL,
#   verbose_mode = FALSE,
#   compress_rasters = FALSE
# )
```

```{r}
my_dir8 <- glue('08_sobelmask20/{basename(my_site_dir)}')
fs::dir_create(my_dir8)

wb_filename1 <- glue('{my_dir8}/sobelmask20.tif')
riv_sobel_mask20 %>% terra::writeRaster(wb_filename1, overwrite = TRUE)

my_dir9 <- glue('09_wb_patches/{basename(my_site_dir)}')
fs::dir_create(my_dir9)

wb_filename2 <- glue('{my_dir9}/patches.tif')

wbt_clump(
  input = wb_filename1,
  output = wb_filename2,
  diag = TRUE,
  verbose_mode = TRUE,
  compress_rasters = TRUE
)
```

```{r}
# riv_sobel_mask20_patches <- terra::patches(riv_sobel_mask20, directions = 8)
riv_sobel_mask20_patches <- terra::rast(wb_filename2)


wb_dir9c <- glue('{data_dir}/centerlines/09_patches/{basename(my_site_dir)}')
fs::dir_create(wb_dir9c)
riv_sobel_mask20_patches %>% terra::writeRaster(glue('{wb_dir9c}/patches.tif'))

# gg3 <- ggplot() +
#   geom_stars(data = st_as_stars(riv_sobel_mask20_patches)) +
#   scale_fill_viridis_c() +
#   theme_void()
# # 
# ggsave(glue('{my_dir9}/{basename(my_site_dir)}-patches.pdf'), gg3)

```


```{r}
# convert to points
riv_cent_sv <- terra::as.points(riv_sobel_mask01)
riv_cent_sf <- riv_cent_sv %>% st_as_sf() %>% tibble::rowid_to_column('pt_id')
# Save points
# plot(riv_cent_sf)
pts_dir <- glue('{data_dir}/centerlines/09_centerpts/{basename(my_site_dir)}')
fs::dir_create(pts_dir)
riv_cent_sf %>% st_write(glue('{pts_dir}/{basename(my_site_dir)}_mask01pts.shp'))
```
 
```{r}
library(leaflet)
library(htmltools)

patches4326 <- terra::project(riv_sobel_mask20_patches, 'EPSG:4326')
minmax(patches4326)
patches_r <- raster::raster(patches4326)
my_pal <- colorNumeric('viridis', na.color = "#00000000", domain = 1:minmax(patches4326)[2])

ll <- leaflet() %>%
  addTiles(group = 'bg') %>%
  addProviderTiles(providers$Esri.WorldImagery, group = 'esri') %>%
  addRasterImage(x = patches_r, colors = my_pal, maxBytes = 4e6, group = 'patches') %>%
  addMarkers(data = st_transform(riv_cent_sf, 4326),   
             clusterOptions = markerClusterOptions(),
             group = 'pts', popup = ~glue('point {pt_id}')) %>%
  addLayersControl(overlayGroups = c('bg','pts', 'esri', 'patches'))
ll
# ll %>% htmltools::save_html(file = glue('10_leaflets/leaflet-{basename(my_site_dir)}.html'))
```

# after points identified

```{r}
seg_df <- readxl::read_excel(glue('11_segments/segments-{basename(my_site_dir)}.xlsx'))
padn <- nchar(nrow(riv_cent_sf))
seg_pts <- riv_cent_sf %>% dplyr::filter(pt_id %in% c(seg_df$pt_start, seg_df$pt_end)) %>%
  dplyr::mutate(seg_pt_id = glue('{basename(my_site_dir)}-pt_{stringr::str_pad(pt_id, width = padn, side = "left", pad = "0")}'))

pts_dir <- glue('{data_dir}/centerlines/09_centerpts/{basename(my_site_dir)}')
seg_pts %>% st_write(glue('{pts_dir}/{basename(my_site_dir)}_seg-pts.shp'))
```



```{r}
# bb <- ggplot() +
#     geom_stars(data = st_as_stars(my_mndwi)) +
#        colorspace::scale_fill_continuous_diverging(palette = 'Cork', mid = 0, rev = TRUE) +
#      geom_sf(data = riv_smooth, fill = NA, col = NA) +
#     theme_void() +
#    theme(legend.position = 'bottom') +
#   ggtitle(glue('{my_reflectance_id} {my_aop_yr}'))
# 
# cc <- ggplot() +
#     geom_stars(data = st_as_stars(my_mndwi)) +
#        colorspace::scale_fill_continuous_diverging(palette = 'Cork', mid = 0, rev = TRUE) +
#      geom_sf(data = riv_smooth, fill = NA, col = 'black') +
#     theme_void() +
#    theme(legend.position = 'bottom') +
#    ggtitle('')
# 
# cowplot::plot_grid(bb, cc, nrow = 1)
# 
# ggsave(glue('figs/{my_reflectance_id}-{my_aop_yr}.png'), width = 12)

```
