==> 02-watermask-rivgroups.Rmd <==
---
title: "01-watermask-mosaic"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
data_dir <- 'S:/users/khondula/DATA'

fs::dir_ls(glue('{data_dir}/centerlines/01_watermask-imgs')) %>% basename()
fs::dir_ls(glue('{data_dir}/centerlines/01_mNDWI')) %>% basename()

my_site_yr <- fs::dir_ls(glue('{data_dir}/centerlines/01_mNDWI')) %>% basename() %>%
  grep('CLBJ', x = ., value = TRUE)
```

# MNDWI

```{r}
mask_dir <- glue('{data_dir}/centerlines/01_mNDWI/{my_site_yr}')
# mask_dir <- glue('{data_dir}/centerlines/03_watermask-data/{my_site_yr}')
mask_files <- fs::dir_ls(mask_dir, glob = glue('*.tif')) %>% as.character()
```

```{r}
rivercheck_dir <- glue('{data_dir}/centerlines/riverchecks')
check_df <- readr::read_csv(glue('{rivercheck_dir}/{my_site_yr}-rivercheck-notes.csv')) %>% dplyr::filter(river_check)
if(!'group' %in% names(check_df)){check_df <- check_df %>% dplyr::rename(group = notes)}
sort(unique(check_df$group))
```

```{r}
my_group <- sort(unique(check_df$group))[2] # NA if no groups
my_group
my_group_id <- 'clbj2'
```

```{r}
group_df <- check_df %>% dplyr::filter(group %in% my_group)
my_river_fileids <- group_df$refl_id
idz <- my_river_fileids %>% purrr::map_int(~grep(pattern = .x, x = mask_files))
river_site_files <- mask_files[idz]
length(river_site_files)
```

# mosaic with gdal

```{r}
gdalcommand <- glue::glue('gdalbuildvrt tmp/mosaic.vrt {glue_collapse(river_site_files, sep = " ")}')
system(gdalcommand)
gdalcommand2 <- glue('gdal_translate -of GTiff tmp/mosaic.vrt tmp/mosaic-{my_site_yr}.tif')
system(gdalcommand2)

m <- terra::rast(glue('tmp/mosaic-{my_site_yr}.tif'))
m %>% writeRaster(glue('{data_dir}/centerlines/03_watermask-rivgroups/{my_site_yr}-{my_group_id}-MNDWI.tif'), overwrite = TRUE)
plot(m)

my_bbox <- st_bbox(m) %>% st_as_sfc() %>% st_as_sf()
m5 <- terra::aggregate(m, fact = 5, na.rm = TRUE)
gg1 <- ggplot() +
  geom_stars(data = st_as_stars(m5)) +
  colorspace::scale_fill_continuous_sequential(palette = 'Plasma', rev = TRUE) +
  geom_sf(data = my_bbox, fill = NA, col = 'black') +
  theme_void() +
  ggtitle(glue('{my_site_yr} mNDWI')) +
  theme(legend.position = 'none')
  
plot_dir <- glue('{data_dir}/centerlines/03_rivmosaic-imgs/{my_site_yr}')
fs::dir_create(plot_dir)
ggsave(glue('{plot_dir}/{my_site_yr}-{my_group_id}-mNDWI.pdf'), gg1, height = 12, width = 8)

rcl_matrix <- matrix(c(-Inf, 0, 0, 0, Inf, 1), ncol = 3, byrow = TRUE)
m01 <- terra::classify(m, rcl_matrix, include.lowest = FALSE)
plot(m01)
m01 %>% writeRaster(glue('{data_dir}/centerlines/03_watermask-rivgroups/{my_site_yr}-{my_group_id}-MNDWI_01.tif'), overwrite = TRUE)
```

# AWEI

```{r}
mask_dir <- glue('{data_dir}/centerlines/01_AWEI/{my_site_yr}')
mask_files <- fs::dir_ls(mask_dir, glob = glue('*.tif')) %>% as.character()

group_df <- check_df %>% dplyr::filter(group %in% my_group)
my_river_fileids <- group_df$refl_id
idz <- my_river_fileids %>% purrr::map_int(~grep(pattern = .x, x = mask_files))
river_site_files <- mask_files[idz]
length(river_site_files)

# mosaic with gdal

gdalcommand <- glue::glue('gdalbuildvrt tmp/mosaic.vrt {glue_collapse(river_site_files, sep = " ")}')
system(gdalcommand)
gdalcommand2 <- glue('gdal_translate -of GTiff tmp/mosaic.vrt tmp/mosaic-{my_site_yr}.tif')
system(gdalcommand2)

m <- terra::rast(glue('tmp/mosaic-{my_site_yr}.tif'))
m %>% writeRaster(glue('{data_dir}/centerlines/03_watermask-rivgroups/{my_site_yr}-{my_group_id}-AWEI.tif'), overwrite = TRUE)
plot(m)

my_bbox <- st_bbox(m) %>% st_as_sfc() %>% st_as_sf()
m5 <- terra::aggregate(m, fact = 5, na.rm = TRUE)
gg1 <- ggplot() +
  geom_stars(data = st_as_stars(m5)) +
  colorspace::scale_fill_continuous_sequential(palette = 'Plasma', rev = TRUE) +
  geom_sf(data = my_bbox, fill = NA, col = 'black') +
  theme_void() +
  ggtitle(glue('{my_site_yr} AWEI')) +
  theme(legend.position = 'none')
  
plot_dir <- glue('{data_dir}/centerlines/03_rivmosaic-imgs/{my_site_yr}')
fs::dir_create(plot_dir)
ggsave(glue('{plot_dir}/{my_site_yr}-{my_group_id}-AWEI.pdf'), gg1, height = 12, width = 12) # adjust fig size if needed
```
